+++ packages/manager/CHANGELOG.md
@@ -2,7 +2,6 @@
 
 ## 1.0.0 (2026-03-27)
 
-
 ### Features
 
-* **tui:** add manager and client ([0e01cab](https://github.com/ThomasRumas/local-llm-manager/commit/0e01cab6de462d481b80b0f93d2b913c25c504f3))
+- **tui:** add manager and client ([0e01cab](https://github.com/ThomasRumas/local-llm-manager/commit/0e01cab6de462d481b80b0f93d2b913c25c504f3))